+++ brokers/theia/broker_test.go
@@ -24,14 +24,14 @@ import (
 
 	tests "github.com/eclipse/che-plugin-broker/brokers/test"
 	cmock "github.com/eclipse/che-plugin-broker/common/mocks"
-	fmock "github.com/eclipse/che-plugin-broker/files/mocks"
 	"github.com/eclipse/che-plugin-broker/model"
 	"github.com/eclipse/che-plugin-broker/storage"
+	umock "github.com/eclipse/che-plugin-broker/utils/mocks"
 )
 
 var (
 	bMock      = &cmock.Broker{}
-	uMock      = &fmock.IoUtil{}
+	uMock      = &umock.IoUtil{}
 	randMock   = &cmock.Random{}
 	mockBroker = &Broker{
 		Broker:  bMock,
